+++ src/components/ColorGraph/Chart/RenderStrategy/WorkerPool/CanvasWorkerPool.ts
@@ -1,9 +1,9 @@
-import { DrawChartProps, WorkerObj } from 'worker/paintWorker'
 import { Channel } from 'shared/types'
+import { DrawChartProps, WorkerObj } from './worker/paintWorker'
 
 // @ts-ignore Module not found
 // eslint-disable-next-line import/no-webpack-loader-syntax
-import PaintWorker from 'worker/paintWorker?worker&inline'
+import PaintWorker from './worker/paintWorker?worker&inline'
 import * as Comlink from 'comlink'
 
 export type PaintResult = ImageBitmap | ImageData
