+++ src/components/ColorGraph/Chart/RenderStrategy/WorkerPool/worker/Pixels.ts
