--- src/worker/worker.ts
@@ -1,4 +1,0 @@
-import * as Comlink from 'comlink'
-import { obj } from './paintWorker'
-
-Comlink.expose(obj)
